+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Flutter Widgetz
 
+## 0.0.2
+
+- Add `CustomBody` widget.
+- Allow more user params with `CustomSearchBar`.
+- Set default icon on `CustomInputDate`.
+
 ## 0.0.1
 
 - Initial developer preview.